[IFB-730] feat: 변경된  DB 구조 적용

+++ src/main/java/kr/okku/server/dto/controller/review/PickDto.java
@@ -8,6 +8,6 @@ public class PickDto {
     private String id;
     private String image;
     private String name;
-    private double price;
+    private Integer price;
     private String url;
 }
